+++ admin/tabs/server/ehcache/EhCachePanel.js
@@ -5,7 +5,7 @@
 * Copyright © 2021 Extremely Heavy Industries Inc.
 */
 import {grid, gridCountLabel} from '@xh/hoist/cmp/grid';
-import {filler} from '@xh/hoist/cmp/layout';
+import {filler, span} from '@xh/hoist/cmp/layout';
 import {storeFilterField} from '@xh/hoist/cmp/store';
 import {creates, hoistCmp} from '@xh/hoist/core';
 import {button, exportButton} from '@xh/hoist/desktop/cmp/button';
@@ -20,15 +20,22 @@ export const ehCachePanel = hoistCmp.factory({
         return panel({
             mask: 'onLoad',
             tbar: [
+                Icon.info(),
+                span({
+                    item: 'Hibernate (Ehcache) caches for server-side domain objects',
+                    className: 'xh-bold'
+                }),
+                filler(),
                 button({
                     icon: Icon.reset(),
                     text: 'Clear All',
                     intent: 'danger',
                     onClick: () => model.clearAllAsync()
                 }),
-                filler(),
+                '-',
                 gridCountLabel({unit: 'cache'}),
-                storeFilterField(),
+                '-',
+                storeFilterField({matchMode: 'any'}),
                 exportButton()
             ],
             item: grid()